requirements.txt
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.2
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
asttokens==3.0.0
attrs==20.3.0
Babel==2.10.1
blinker==1.4
cachelib==0.7.0
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
comm==0.2.3
commonmark==0.9.1
connexion==2.13.1
cron_descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
exceptiongroup==1.3.0
executing==2.2.0
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
ipykernel==6.30.1
ipython==8.18.1
itsdangerous==1.1.0
jedi==0.19.2
Jinja2==3.0.3
jsonschema==4.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.12
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.0.0
prison==0.2.1
prompt_toolkit==3.0.51
psutil==5.9.1
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.1
pyspark==3.4.4
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==27.0.1
requests==2.27.1
rfc3986==1.5.0
rich==12.4.4
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
stack-data==0.6.3
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
unicodecsv==0.14.1
urllib3==1.26.9
wcwidth==0.2.13
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.2
apache-airflow-providers-amazon==7.3.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==20.3.0
Babel==2.10.1
beautifulsoup4==4.13.5
blinker==1.4
boto3==1.40.19
botocore==1.40.19
cachelib==0.7.0
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
comm==0.2.3
commonmark==0.9.1
connexion==2.13.1
cron_descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
exceptiongroup==1.3.0
executing==2.2.0
findspark==2.0.1
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
ipykernel==6.30.1
ipython==8.18.1
itsdangerous==1.1.0
jedi==0.19.2
Jinja2==3.0.3
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.12
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==5.4.0
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mypy-boto3-appflow==1.40.17
mypy-boto3-rds==1.40.16
mypy-boto3-redshift-data==1.40.0
nest-asyncio==1.6.0
numpy==2.0.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.0.0
ply==3.11
prison==0.2.1
prompt_toolkit==3.0.51
psutil==5.9.1
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==21.0.0
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.1
pyspark==3.4.4
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==27.0.1
redshift-connector==2.1.8
requests==2.27.1
rfc3986==1.5.0
rich==12.4.4
s3transfer==0.13.1
scramp==1.4.6
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.8
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.38.2
sqlparse==0.5.3
stack-data==0.6.3
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
unicodecsv==0.14.1
urllib3==1.26.9
watchtower==2.0.1
wcwidth==0.2.13
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
